move coordinate header downcasing to Clusters Class
