+++ tests/test_annotations.py
@@ -26,6 +26,7 @@ import numpy as np
 
 sys.path.append("../ingest")
 from annotations import Annotations
+from clusters import Clusters
 
 
 class TestAnnotations(unittest.TestCase):
@@ -66,42 +67,33 @@ class TestAnnotations(unittest.TestCase):
                     "Group annotations must be string values",
                 )
 
-    def test_non_coordinate_case_integrity(self):
-        # if headers of this file change, the reference for the assert needs to be updated
-        self.df.preprocess()
-        assert self.df.headers == [
-            'NAME',
-            'x',
-            'y',
-            'z',
-            'CLUSTER',
-            'SUBCLUSTER',
-            '1',
-        ], 'preprocess should only downcase coordinate header columns'
-
     def test_merge_df(self):
-        self.df.preprocess()
+        cluster = Clusters(
+            '../tests/data/test_1k_cluster_data.csv',
+            'dec0dedfeed1111111111111',
+            'addedfeed000000000000000',
+            'testCluster',
+        )
         cell_metadata_df = Annotations(
             self.CELL_METADATA_PATH,
             ['text/csv', 'text/plain', 'text/tab-separated-values'],
         )
         cell_metadata_df.preprocess()
-        self.df.determine_coordinates_and_cell_names()
         cell_names_cell_metadata_df = np.asarray(cell_metadata_df.file['NAME'])
-        cell_names_cluster_df = np.asarray(self.df.file['NAME'])
+        cell_names_cluster_df = np.asarray(cluster.file['NAME'])
         # Cell names found in both cluster and metadata files
         common_cell_names = cell_names_cluster_df[
             np.isin(cell_names_cluster_df, cell_names_cell_metadata_df)
         ]
         print(f'common cell names: {common_cell_names}')
         # Perform merge
-        print(self.df.file[['NAME', 'x', 'y', 'z']])
-        self.df.merge_df(self.df.file[['NAME', 'x', 'y', 'z']], cell_metadata_df.file)
+        print(cluster.file[['NAME', 'x', 'y', 'z']])
+        cluster.merge_df(cluster.file[['NAME', 'x', 'y', 'z']], cell_metadata_df.file)
 
         # Ensure ONLY common cell names found in cell metadata file and cluster file
         # are in the newly merged df
         result = all(
-            cell[0] in common_cell_names for cell in self.df.file['NAME'].values
+            cell[0] in common_cell_names for cell in cluster.file['NAME'].values
         )
         self.assertTrue(
             result,
